backend/requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-fsm==3.0.0
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# AWS S3
boto3==1.34.14
django-storages==1.14.2

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.1.0
WeasyPrint==60.1

# ORCID & OAuth
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0
whitenoise==6.6.0

# Development
python-decouple==3.8
django-extensions==3.2.3
ipython==8.18.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0

